payments/views.py
from django.shortcuts import render
from rest_framework.decorators import api_view, permission_classes
from rest_framework.permissions import IsAuthenticated
from rest_framework.response import Response
from rest_framework import status

# 临时占位视图，后续会完善

@api_view(['POST'])
@permission_classes([IsAuthenticated])
def payment_create(request):
    """创建支付API"""
    return Response({'message': '创建支付功能待实现'})

@api_view(['GET'])
@permission_classes([IsAuthenticated])
def payment_list(request):
    """支付列表API"""
    return Response({'message': '支付列表功能待实现'})

@api_view(['GET'])
@permission_classes([IsAuthenticated])
def payment_detail(request, payment_id):
    """支付详情API"""
    return Response({'message': f'支付{payment_id}详情功能待实现'})

@api_view(['POST'])
@permission_classes([IsAuthenticated])
def payment_confirm(request, payment_id):
    """确认支付API"""
    return Response({'message': f'确认支付{payment_id}功能待实现'})

@api_view(['POST'])
@permission_classes([IsAuthenticated])
def refund_create(request):
    """创建退款API"""
    return Response({'message': '创建退款功能待实现'})

@api_view(['GET'])
@permission_classes([IsAuthenticated])
def refund_list(request):
    """退款列表API"""
    return Response({'message': '退款列表功能待实现'})

@api_view(['GET'])
@permission_classes([IsAuthenticated])
def refund_detail(request, refund_id):
    """退款详情API"""
    return Response({'message': f'退款{refund_id}详情功能待实现'})

@api_view(['POST'])
@permission_classes([IsAuthenticated])
def refund_approve(request, refund_id):
    """批准退款API"""
    return Response({'message': f'批准退款{refund_id}功能待实现'})

@api_view(['GET'])
@permission_classes([IsAuthenticated])
def user_account(request):
    """用户账户API"""
    return Response({'message': '用户账户功能待实现'})

@api_view(['POST'])
@permission_classes([IsAuthenticated])
def account_recharge(request):
    """账户充值API"""
    return Response({'message': '账户充值功能待实现'})

@api_view(['GET'])
@permission_classes([IsAuthenticated])
def account_transactions(request):
    """账户交易记录API"""
    return Response({'message': '账户交易记录功能待实现'})

@api_view(['POST'])
@permission_classes([IsAuthenticated])
def invoice_create(request):
    """创建发票API"""
    return Response({'message': '创建发票功能待实现'})

@api_view(['GET'])
@permission_classes([IsAuthenticated])
def invoice_list(request):
    """发票列表API"""
    return Response({'message': '发票列表功能待实现'})

@api_view(['GET'])
@permission_classes([IsAuthenticated])
def invoice_detail(request, invoice_number):
    """发票详情API"""
    return Response({'message': f'发票{invoice_number}详情功能待实现'})

@api_view(['GET'])
@permission_classes([IsAuthenticated])
def payment_methods(request):
    """支付方式API"""
    return Response({'message': '支付方式功能待实现'})